requirements.txt
deep-translator
gradio
mflux
python-dotenv
torch
torchvision
mlx
mlx-vlm
customtkinter
Pillow
huggingface_hub
numpy
requests
